+++ codewof/users/views.py
@@ -37,6 +37,9 @@ from users.models import Group, Membership, GroupRole, Invitation
 
 from programming.codewof_utils import get_questions_answered_in_past_month, backdate_user
 
+from users.mixins import AdminRequiredMixin, AdminOrMemberRequiredMixin, SufficientAdminsMixin, \
+    RequestUserIsMembershipUserMixin
+
 User = get_user_model()
 
 logger = logging.getLogger(__name__)
@@ -239,29 +242,6 @@ class GroupCreateView(LoginRequiredMixin, CreateView):
         return response
 
 
-class AdminRequiredMixin:
-    """Mixin for checking the user is an Admin of the Group."""
-
-    def dispatch(self, request, *args, **kwargs):
-        """Dispatch for AdminRequiredMixin."""
-        admin_role = GroupRole.objects.get(name="Admin")
-        if Membership.objects.all().filter(user=self.request.user, group=self.get_object(), role=admin_role):
-            return super().dispatch(request, *args, **kwargs)
-        else:
-            raise PermissionDenied()
-
-
-class AdminOrMemberRequiredMixin:
-    """Mixin for checking the user is an Admin or Member of the Group."""
-
-    def dispatch(self, request, *args, **kwargs):
-        """Dispatch for AdminOrMemberRequiredMixin."""
-        if Membership.objects.all().filter(user=self.request.user, group=self.get_object()):
-            return super().dispatch(request, *args, **kwargs)
-        else:
-            raise PermissionDenied()
-
-
 class GroupUpdateView(LoginRequiredMixin, AdminRequiredMixin, UpdateView):
     """View for updating a group."""
 
@@ -372,32 +352,6 @@ def update_memberships(request, pk, group):
         raise Exception("Must have at least one Admin in the group.")
 
     return HttpResponse()
-
-
-class SufficientAdminsMixin:
-    """Mixin for checking there will be enough Admins if deleting an Admin Membership."""
-
-    def dispatch(self, request, *args, **kwargs):
-        """Dispatch for SufficientAdminsMixin."""
-        membership = self.get_object()
-        admin_role = GroupRole.objects.get(name='Admin')
-        if len(Membership.objects.all().filter(group=membership.group,
-                                               role=admin_role)) > 1 or membership.role != admin_role:
-            return super().dispatch(request, *args, **kwargs)
-        else:
-            raise Exception("A Group must have at least one Admin.")
-
-
-class RequestUserIsMembershipUserMixin:
-    """Mixin for checking the user making the request is the user of the Membership."""
-
-    def dispatch(self, request, *args, **kwargs):
-        """Dispatch for RequestUserIsMembershipUserMixin."""
-        membership = self.get_object()
-        if request.user == membership.user:
-            return super().dispatch(request, *args, **kwargs)
-        else:
-            raise PermissionDenied()
 
 
 class MembershipDeleteView(LoginRequiredMixin, RequestUserIsMembershipUserMixin, SufficientAdminsMixin, DeleteView):
